python-115-share-link-webdav/requirements.txt
cheroot
pycryptodome
pyyaml
qrcode
requests
watchdog
WsgiDAV